gemini/webgem/python/fastapi/requirements.txt
fastapi[standard]
mistune
pillow
google-genai
